Improve parts availability for EVA suit related malfunctions
r5075
2019-10-06

## CHANGE
1. Modify getTimeString() for the time compression label
   in ClockUtils and String.format() for the time ratio in
   TimeWindow.
   - Remove the 2 decimal places.
2. Add the overriding getPartsNeededForTrip() method in
   Mining and Exploration.
   - Note : this method has already been used in only
     CollectionResourceMission. Should it be used in all
	 mission involing EVA ?
3. Modify getPartsNeededForTrip() to match only the EVA
   suit parts when calculating the repair parts for EVA
   suit.
4. Remove the possibility of bringing any '3d printer' in
   getPartsNeededForTrip() in VehicleMission.
   - Note: will investigate why in future.
5. Add more parts that may fail in case of EVA suit related
   malfunctions in malfunctions.mxl.

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/time/ClockUtils.java
@@ -199,7 +199,7 @@ public class ClockUtils implements Serializable {
 			b.append(ZERO_MINUTES);
 		}
 
-		b.append(String.format("%05.2f", secs) + SECONDS);
+		b.append(String.format("%02d", (int)secs) + SECONDS);
 
 		return b.toString();
 	}
